pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "discord-alerts-kit"
version = "0.1.0"
authors = [
  { name="jackfrost_13" }
]
description = "A simple Python library to build Discord alert bots in minutes"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "discord.py>=2.0.0",
    "pyyaml>=6.0"
]

[project.urls]
Homepage = "https://github.com/jackfrost_13/discord-alerts-kit"